+++ CHANGELOG.rst
@@ -2,6 +2,38 @@
 What's New
 ==========
 
+climpred v2.1.4 (2021-XX-XX)
+============================
+
+Breaking changes
+----------------
+
+New Features
+------------
+- Allow ``hours``, ``minutes`` and ``seconds`` as ``lead.attrs['units']``.
+  (:issue:`404`, :pr:`603`) `Aaron Spring`_.
+- Allow to set ``seasonality`` via :py:class:`~climpred.options.set_options` to specify
+  how to group in ``verify(reference='climatology'`` or in
+  :py:meth:`~climpred.classes.HindcastEnsemble.remove_bias`.
+  (:issue:`529`, :pr:`603`) `Aaron Spring`_.
+- Allow ``weekofyear`` via ``datetime`` in
+  :py:meth:`~climpred.classes.HindcastEnsemble.remove_bias`, but not yet implemented in
+  ``verify(reference='climatology')``. (:issue:`529`, :pr:`603`) `Aaron Spring`_.
+
+Internals/Minor Fixes
+---------------------
+- Add weekly upstream CI, which raises issues for failures. Adapted from ``xarray``.
+  Manually trigger by ``git commit -m '[test-upstream]'``. Skip climpred_testing CI by
+  ``git commit -m '[skip-ci]'``
+  (:issue:`518`, :pr:`596`) `Aaron Spring`_.
+- Add `documentation page about publicly available initialized datasets and
+  corresponding climpred examples <initialized-datasets.html>`_.
+  (:issue:`510`, :issue:`561`, :pr:`600`) `Aaron Spring`_.
+- Add `GEFS example <examples/NWP/NWP_GEFS_6h_forecasts.html>`_ for numerical weather
+  prediction. (:issue:`602`, :pr:`603`) `Aaron Spring`_.
+- Add subseasonal `daily ECMWF example <examples/daily-S2S-ECMWF.html>`_ using
+  `climetlab <https://github.com/ecmwf-lab/climetlab-s2s-ai-challenge>`_ to access
+  hindcasts from ECMWF cloud.  (:issue:`587`, :pr:`603`) `Aaron Spring`_.
 
 climpred v2.1.3 (2021-03-23)
 ============================
@@ -56,6 +88,7 @@ Internals/Minor Fixes
   Find further examples in the ``examples`` folder.
   (:issue:`549`, :pr:`578`) `Aaron Spring`_.
 - Rename branch ``master`` to ``main``. (:pr:`579`) `Aaron Spring`_.
+
 
 climpred v2.1.2 (2021-01-22)
 ============================